x/testhellochain/types/message_create_venue_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"testhellochain/testutil/sample"
)

func TestMsgCreateVenue_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateVenue
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateVenue{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateVenue{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
